requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
pyTelegramBotAPI==4.2.2
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
urllib3==1.26.7
Werkzeug==2.0.2
